requirements.txt
https://download.pytorch.org/whl/cu100/torch-1.0.0-cp37-cp37m-win_amd64.whl
torchvision
numpy
opencv-python
pyqt5